Backend/requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.2.10
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
bootstrap4==0.1.0
cffi==1.14.0
click==7.1.2
colorama==0.4.3
decorator==4.4.2
distlib==0.3.0
Django==3.0.7
django-bootstrap4==2.0.1
django-cors-headers==3.4.0
django-filter==2.3.0
django-phone-field==1.8.1
djangorestframework==3.11.0
filelock==3.0.12
Flask==1.1.2
ipython==7.15.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
MarkupSafe==1.1.1
misaka==2.1.1
parso==0.7.0
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==3.0.5
psycopg2==2.8.5
pycparser==2.20
Pygments==2.6.1
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
traitlets==4.3.3
virtualenv==20.0.21
wcwidth==0.1.9
Werkzeug==1.0.1
